refactor: 대기 소켓 핸들러 파일 위치 변경
