+++ src/v1/sockets/waiting/waiting.namespace.ts
@@ -2,7 +2,7 @@ import { Namespace, Socket } from 'socket.io';
 import { socketMiddleware } from '../utils/middleware.js';
 import { socketErrorHandler } from '../utils/errorHandler.js';
 import { autoJoinSchemaType } from './schemas/auto-game.schema.js';
-import WaitingSocketHandler from './waiting.socket.handler.js';
+import WaitingSocketHandler from './handlers/waiting.socket.handler.js';
 import SocketCache from '../../storage/cache/socket.cache.js';
 import { SOCKET_EVENTS } from './waiting.event.js';
 import { FastifyBaseLogger } from 'fastify';
